# internal/httpapi/openapi.yaml
openapi: 3.1.0
info:
  title: Open BIN API
  version: 1.0.0
  description: |
    A free, keyless BIN/IIN lookup API backed by automatically synchronized,
    community-maintained GitHub datasets. The application has no per-user rate limit.
    Only 6 to 8 digit issuer identifiers are accepted; never send a full card number.
  license:
    name: MIT
paths:
  /v1/bin/{iin}:
    get:
      summary: Look up a BIN/IIN
      operationId: lookupIIN
      parameters:
        - name: iin
          in: path
          required: true
          description: The first 6 to 8 digits of a payment card number.
          schema:
            type: string
            pattern: '^[0-9]{6,8}$'
      responses:
        '200':
          description: The most specific record covering this IIN.
          headers:
            ETag:
              schema: { type: string }
            Cache-Control:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lookup' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /{iin}:
    get:
      summary: Compatibility lookup route
      description: Alias of GET /v1/bin/{iin} for clients migrating from binlist.net.
      operationId: lookupIINLegacy
      parameters:
        - name: iin
          in: path
          required: true
          schema: { type: string, pattern: '^[0-9]{6,8}$' }
      responses:
        '200':
          description: Lookup result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lookup' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/bins:lookup:
    post:
      summary: Look up as many as 1,000 IINs in one request
      operationId: lookupIINBatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bins]
              properties:
                bins:
                  type: array
                  minItems: 1
                  maxItems: 1000
                  items: { type: string, pattern: '^[0-9]{6,8}$' }
      responses:
        '200':
          description: Ordered lookup results
          content:
            application/json:
              schema:
                type: object
                required: [results]
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      required: [iin, found, result]
                      properties:
                        iin: { type: string }
                        found: { type: boolean }
                        result:
                          oneOf:
                            - { $ref: '#/components/schemas/Lookup' }
                            - { type: 'null' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /v1/stats:
    get:
      summary: Data freshness and source status
      operationId: stats
      responses:
        '200':
          description: Import status
          content:
            application/json:
              schema:
                type: object
                required: [records, sources]
                properties:
                  records: { type: integer, format: int64 }
                  last_synced_at: { type: [string, 'null'], format: date-time }
                  sources:
                    type: array
                    items: { $ref: '#/components/schemas/SourceStatus' }
  /healthz:
    get:
      summary: Liveness check
      responses:
        '200': { description: Process is alive }
  /readyz:
    get:
      summary: Readiness check
      responses:
        '200': { description: At least one dataset is queryable }
        '503': { description: Initial synchronization has not completed }
components:
  schemas:
    Lookup:
      type: object
      required: [iin, match, number, scheme, brand, type, level, prepaid, country, bank, source]
      properties:
        iin: { type: string, examples: ['45717360'] }
        match:
          type: object
          required: [start, end, length]
          properties:
            start: { type: string }
            end: { type: string }
            length: { type: integer }
        number:
          type: object
          required: [length, luhn]
          properties:
            length: { type: [integer, 'null'] }
            luhn: { type: [boolean, 'null'] }
        scheme: { type: string, examples: [visa] }
        brand: { type: string }
        type: { type: string, examples: [debit] }
        level: { type: string, examples: [classic] }
        prepaid: { type: [boolean, 'null'] }
        country:
          type: object
          required: [alpha2, alpha3, name, emoji, currency, latitude, longitude]
          properties:
            alpha2: { type: string }
            alpha3: { type: string }
            name: { type: string }
            emoji: { type: string }
            currency: { type: string }
            latitude: { type: [number, 'null'] }
            longitude: { type: [number, 'null'] }
        bank:
          type: object
          required: [name, url, phone, city, logo]
          properties:
            name: { type: string }
            url: { type: string }
            phone: { type: string }
            city: { type: string }
            logo: { type: string }
        source:
          type: object
          required: [id, repository, commit, synced_at]
          properties:
            id: { type: string }
            repository: { type: string }
            commit: { type: string }
            synced_at: { type: string, format: date-time }
    SourceStatus:
      type: object
      properties:
        id: { type: string }
        repository: { type: string }
        branch: { type: string }
        path: { type: string }
        priority: { type: integer }
        status: { type: string, enum: [pending, syncing, ready, error] }
        current_commit: { type: string }
        record_count: { type: integer, format: int64 }
        skipped_rows: { type: integer, format: int64 }
        last_checked_at: { type: [string, 'null'], format: date-time }
        last_synced_at: { type: [string, 'null'], format: date-time }
        last_error: { type: string }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
  responses:
    BadRequest:
      description: Invalid IIN or request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: No community record covers this IIN
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }